refactor: extract upload logic into uploadController and simplify uploadRoutes
